Data Collection/requirements.txt
aiohttp==3.9.3
beanie==1.25.0
duckduckgo-search==4.4.3
fastapi==0.110.0
ftfy==6.1.3
motor==3.3.2
nltk==3.8.1
numpy==1.26.4
openai==1.12.0
pandas==2.2.1
pydantic==2.6.3
pymongo==4.6.2
python-dotenv==1.0.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
spacy==3.7.4
tortoise-orm==0.20.0
urllib3==2.2.1

# Additional spaCy model
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
